# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Example",
        "description": "This is a sample API documentation for the Laravel project.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Retrieve a list of contacts",
                "operationId": "e670331f5da6af785062a857ec92def2",
                "responses": {
                    "200": {
                        "description": "List of contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a new contact",
                "operationId": "77b53a0b0cb97bf04eafa192e4cbac4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get a specific contact",
                "operationId": "4c4e002507e3a53acc022b210aad20ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Update a specific contact",
                "operationId": "adc814a54ff3a1fdc424293b19ffe8a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Delete a specific contact",
                "operationId": "bc8b0b91abd89481c321fcdb966dc355",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contact deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "This is a sample message."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ContactRequest": {
                "required": [
                    "name",
                    "phone",
                    "email",
                    "message"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "Hello, this is John."
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Contacts",
            "description": "API Endpoints for Managing Contacts"
        }
    ]
}